git-svn-id: https://taskunifier.svn.sourceforge.net/svnroot/taskunifier/TaskUnifierGui@1707 4cc7d151-2d3f-4153-82be-2c8b9e7ce797

+++ src/com/leclercb/taskunifier/gui/swing/TimerField.java
@@ -49,7 +49,7 @@ public class TimerField extends JPanel {
 		CheckUtils.isNotNull(timer, "Timer cannot be null");
 		this.timer = timer;
 		
-		this.timeSpinner.setValue(this.timer.getTimerValue() / 60);
+		this.timeSpinner.setValue((int) (this.timer.getTimerValue() / 60));
 		this.button.setSelected(this.timer.isStarted());
 	}
